Working prototype after finding createPortal cannot work
`createPortal()` can't work because a solution using this requires
React 17 to render a JSX object generated by React 19, which cannot
be done. It results in the following error (from React 17) because
R17 doesn't recognize the object generated by R19 as a component
that it can render:

```
react-dom.development.js:13231 Uncaught Error: Objects are not valid
as a React child (found: object with keys {$$typeof, type, key, props,
_owner, _store}). If you meant to render a collection of children,
use an array instead.
```

So the only solution/way is to render the R19 component in a separate
root, entirely isolated from the R17 root and render tree.